Moved loadsupport to tests-common
The idea is that the simulator package itself is going to get the least amount of
Hazelcast dependencies.
